+++ services/authentication-service/src/providers/google-oauth2-signup.provider.ts
@@ -3,8 +3,6 @@ import {HttpErrors} from '@loopback/rest';
 import {GoogleSignUpFn} from './types';
 
 export class GoogleOauth2SignupProvider implements Provider<GoogleSignUpFn> {
-  constructor() {}
-
   value(): GoogleSignUpFn {
     return async profile => {
       throw new HttpErrors.NotImplemented(