+++ services/ocr-s3-service/src/controllers/ocr-object.controller.ts
@@ -1,10 +1,11 @@
-import { inject } from '@loopback/core';
-import { repository } from '@loopback/repository';
-import { get, param, response } from '@loopback/rest';
-import { HocrObject } from '../models';
-import { HocrObjectRepository } from '../repositories';
-import { IteratorService, S3HandlerService } from '../services';
-import { GetObjectOutput, IListObject } from '../types';
+import {inject} from '@loopback/core';
+import {repository} from '@loopback/repository';
+import {get, param, response} from '@loopback/rest';
+import {STATUS_CODE} from '@sourceloop/core';
+import {HocrObject} from '../models';
+import {HocrObjectRepository} from '../repositories';
+import {IteratorService, S3HandlerService} from '../services';
+import {GetObjectOutput, IListObject} from '../types';
 
 export class OcrObjectController {
   constructor(
@@ -12,10 +13,10 @@ export class OcrObjectController {
     public hocrObjectRepository: HocrObjectRepository,
     @inject('services.S3HandlerService') public s3Handler: S3HandlerService,
     @inject('services.IteratorService') public iteratorService: IteratorService,
-  ) { }
+  ) {}
 
   @get('/get-contract-hocr/{contract_name}')
-  @response(200, {
+  @response(STATUS_CODE.OK, {
     description: 'User model instance',
   })
   async getHocrFiles(
@@ -52,7 +53,7 @@ export class OcrObjectController {
   }
 
   @get('/contract-images/{contract_name}')
-  @response(200, {
+  @response(STATUS_CODE.OK, {
     description: 'User model instance',
   })
   async getImgFiles(
@@ -74,7 +75,9 @@ export class OcrObjectController {
           );
 
           const imgContent: string = await this.s3Handler.streamToString(data);
+          // sonarignore:start
           const contractData: any = {
+            // sonarignore:end
             contractName: contractName,
             type: 'IMG',
             imgData: imgContent,