+++ services/ocr-s3-service/src/repositories/hocr-object.repository.ts
@@ -1,12 +1,11 @@
 import {inject} from '@loopback/core';
 import {DefaultCrudRepository, juggler} from '@loopback/repository';
-import {HocrObject, HocrObjectRelations} from '../models';
+import {HocrObject} from '../models';
 import {OcrS3DbSourceName} from '../types';
 
 export class HocrObjectRepository extends DefaultCrudRepository<
   HocrObject,
-  typeof HocrObject.prototype.id,
-  HocrObjectRelations
+  typeof HocrObject.prototype.id
 > {
   constructor(
     @inject(`datasources.${OcrS3DbSourceName}`)