+++ services/ocr-s3-service/src/services/s3-handler.service.ts
@@ -1,5 +1,5 @@
-import { injectable, BindingScope, inject } from '@loopback/core';
-import { AWSS3Bindings, S3WithSigner } from 'loopback4-s3';
+import {injectable, BindingScope, inject} from '@loopback/core';
+import {AWSS3Bindings, S3WithSigner} from 'loopback4-s3';
 
 @injectable({
   scope: BindingScope.TRANSIENT,
@@ -7,13 +7,13 @@ import { AWSS3Bindings, S3WithSigner } from 'loopback4-s3';
 export class S3HandlerService {
   constructor(
     @inject(AWSS3Bindings.AwsS3Provider) public s3Client: S3WithSigner,
-  ) { }
+  ) {}
 
   async listObjects(
     bucketName: string | undefined,
     contractName: string | undefined,
   ) {
-    const { Contents } = await this.s3Client.listObjectsV2({
+    const {Contents} = await this.s3Client.listObjectsV2({
       Bucket: bucketName,
       Prefix: contractName,
     });
@@ -25,25 +25,26 @@ export class S3HandlerService {
       Bucket: bucketName,
       Key: key,
     });
-    const { Body } = data;
+    const {Body} = data;
     return Body;
   }
 
-  async get(bucketName: string | undefined, key: string | undefined) {
-    const data = await this.s3Client.getObject({
-      Bucket: bucketName,
+  async get(bucket: string | undefined, key: string | undefined) {
+    const output = await this.s3Client.getObject({
+      Bucket: bucket,
       Key: key,
     });
-    const { Body } = data;
+    const {Body} = output;
     return Body;
   }
-
+  // sonarignore:start
   /* eslint-disable  @typescript-eslint/no-explicit-any */
   async streamToString(stream: any): Promise<string> {
+    // sonarignore:end
     return new Promise((resolve, reject) => {
       const chunks: Uint8Array[] = [];
       /* eslint-disable  @typescript-eslint/no-explicit-any */
-      stream.on('data', (chunk: any) => chunks.push(chunk));
+      stream.on('data', (chunk: any) => chunks.push(chunk)); //NOSONAR
       stream.on('error', reject);
       stream.on('end', () => resolve(Buffer.concat(chunks).toString('utf8')));
     });